feat(plugin)!: drop Plugin.dependencies and version-constraint resolution
BREAKING CHANGE: Plugins no longer declare dependencies. The fields
`Plugin.dependencies` and `PluginDependency` are removed, along with
`PluginDependencyError`, `satisfiesVersion`, and the entire
`src/plugin/semver.ts` utility. `hasPlugin` and `getPluginVersion` now
look up only the calling router (no parent traversal, no descent).

Each plugin still gets its own child router and plugin-name uniqueness
on a single router is still enforced via PluginAlreadyInstalledError.

The dependency-validation system surfaced structural issues with the
plugin install pipeline that the name-based dep model could not solve
without significant redesign (wrapper-bundled deps, multi-mount of
shared routers, hook scope, middleware wrapping). Rather than layer
more mechanism on top, the feature is removed; .agents/plugin-design-
followups.md captures the unresolved design issues for a future pass.

Closes #874 by feature removal: the wrapper-plugin dependency
visibility problem no longer exists because no dependency declarations
exist to validate.
