+++ docs/src/.vitepress/config.mjs
@@ -88,7 +88,6 @@ export default defineConfig({
                     text: 'Plugin Authoring',
                     items: [
                         { text: 'Overview', link: '/guide/plugin-authoring/' },
-                        { text: 'Dependencies', link: '/guide/plugin-authoring/dependencies' },
                         { text: 'Conventions', link: '/guide/plugin-authoring/conventions' },
                     ],
                 },
