+++ src/plugin/error/sub/index.ts
@@ -1,4 +1,3 @@
 export * from './already-installed.ts';
-export * from './dependency.ts';
 export * from './install.ts';
 export * from './not-installed.ts';
