+++ src/plugin/index.ts
@@ -1,4 +1,3 @@
 export * from './error/index.ts';
 export * from './is.ts';
-export * from './semver.ts';
 export * from './types.ts';
